kai_bot/requirements.txt
# ─── Core bot dependencies ───────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0

# ─── FastAPI backend ─────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# ─── AI providers ─────────────────────────────────────────────────
google-generativeai>=0.7.0
# ollama (run Ollama locally — no pip install needed)

# ─── Data / analytics ────────────────────────────────────────────
numpy>=1.26.0
# pandas>=2.0.0   (optional — uncomment for advanced analytics)

# ─── Hedera Blockchain SDK ────────────────────────────────────────
hiero-sdk-python==0.1.9

# ─── Hedera AI Agent Kit ──────────────────────────────────────────
# hedera-agent-kit>=3.2.0    (uncomment if installed separately)
